Optimize WebcamPage.js
I noticed that when running this page, it would keep making network requests for the model, so I split the loading of the graph model to its own useEffect that only loads on initial render.

With that speeding things up, I noticed that the timer was continually being initialized, causing a memory leak.
To fix this, I used a useRef instead of useState to track the interval that needed to be cleared.

I also split the detection logic into its own useCallback to make the code easier to understand.

I was able to set the timer to 60ms, and it ran without issue. However, running it that fast caused many letters to be detected which is not a great user experience.